Revert "Single Responsibility principle"
